✨ add new filter for nullable Money
